adding design

+++ util/file.js
@@ -1,10 +1,13 @@
- const fs = require('fs');
- const deleteFile = (filePath)=>{
-     fs.unlink(filePath, (err)=>{
-         if(err){
-             throw(err);
-         }
-     })
- }
+const fs = require('fs');
 
- exports.deleteFile = deleteFile;
+const deleteFile = (filePath) => {
+    if(fs.existsSync(filePath)){
+        fs.unlink(filePath, (err) => {
+            if (err) {
+                throw (err);
+            }
+        });
+    }
+}
+
+exports.deleteFile = deleteFile;
